Refactor subunit validation process in UnitApprovers request.
The validation logic for subunit_id in the DepartmentUnitApprovers request has been simplified. Before, we were retrieving the full record from the database to check whether it exists or not. This has been optimized to only check the existence of the subunit_id. Thus, the change improves application efficiency by reducing database calls.