Add approver_id in response and refactor arrangeLayer method
The approver_id attribute was added to the API response to include the ID of the approver in the 'DepartmentUnitApprovers' controller. This will aid in quickly identifying approver information on the front-end.

The arrangeLayer method is also refactored to accept an ID parameter rather than extracting it from the request. Instead of getting the department ID from the request, it retrieves it from the subunit with the specified ID. The success response message was adjusted to better reflect the changes made.