Refactor responses and improve filtering in AccountTitleController
This commit includes enhancement of responses in the AccountTitleController, by replacing the use of built-in responses with the ApiResponse from the ApiToolkit. This increases code readability and standardizes the response format.

The changes also allow to filter and paginate the accounts by status property, enhancing user experience by providing ordered and paginated data, and aligning functionality with the request-response cycle approach. Moreover, it includes the 'account_title_code' and 'account_title_name' as column search parameters in the 'AccountTitleFilters' class, which improves the search experience of the user.