Implement filters and API responses in various classes and methods
The commit involves adding filter functionalities in PrinterIP, RoleManagement, and TypeOfRequest classes. A new file each for PrinterIPFilters, RoleManagementFilters, and TypeOfRequestFilters is added. It also replaces the static json responses with ApiResponse trait's method in various methods in TypeOfRequestController, MinorCategoryController, RoleManagementController, and PrinterIPController adding consistency and reducing redundancy. This comes with the necessary adjustment to the logic to make it compatible with the changes. This will improve the search functionality and standardize the API responses.