"Refactor AssetRequest handling and implement AssetRequestHandler trait
This commit simplifies the AssetRequestController by moving most of the logic into a new AssetRequestHandler trait. This makes the code more readable, testable, and re-usable. Additionally, a new route (`update-request/{referenceNumber}`) and controller method (`updateRequest`) have been added to handle update requests. This change was needed to cater for updating specific parts of an AssetRequest, as opposed to updating the entire object."