Replace 'asset_request_id' with 'transaction_number' in asset approval process
Updates were made to the handling of asset approvals. Primarily, 'asset_request_id' was replaced with 'transaction_number' in the 'allowedFilters' and 'columnSearch' arrays in the AssetApprovalFilters file, and also in the 'assetRequest' function in the AssetApproval file. Furthermore,  the asset_request_id validation rule was commented out in CreateAssetApprovalRequest file. The 'remove-request_item' route was also modified to 'void-request.' Lastly, a new file - AssetApprovalHandler was added to declutter the transformation of data in AssetApprovalController file. This change came about to ensure consistent usage of identifiers across the application and to enhance code readability and maintainability.