"Change 'Denied' status term to 'Declined'
Updated the term 'Denied' to 'Declined' across multiple files to provide consistent and more appropriate wording. The term 'Denied' was present in CreateAssetApprovalRequest.php, ApprovedRequestRepository.php, AssetRequestController.php, AssetRequestHandler.php and AssetApprovalController.php. In user feedback, 'Declined' was preferred as it is less harsh and more accurately reflects the status of asset requests in certain scenarios."