Add vladimirlogo to .gitignore exclusion list
This commit includes modification to the storage gitignore file by making an addition of 'vladimirlogo/' to the list of files to not ignore. This is necessary as we transition from the Laravel Framework to our new Vladimir Asset Inventory Management System. Such additions will ensure that this file and its updates will be tracked by git, providing better insight into the branding changes being made to the system.