Add specific logo to .gitignore exceptions
In the storage/app/.gitignore file, a particular image file (VladimirSmally-6ce7819f.png) under the 'vladimirlogo/' directory has been added to the list of exceptions for git to track. This was needed as this specific logo file is crucial for certain features of the app and should be part of the git repository ensuring consistent functionality across all deployments.