Add activity logging to AssetApproval model
A relation method 'activityLog' has been added to the 'AssetApproval' model. This was done by importing the 'Activity' model from the Spatie framework and creating a one-to-many relationship between AssetApproval and Activity. The purpose of this change is to provide a way to easily log user activities related to asset approvals. The logging functionality will be useful for future audits and debugging tasks. Note, previous version of activityLog method, using 'latest' scope, is commented out for further consideration.