Add quantity sum to AssetRequestHandler
This update introduces a quantity sum to the AssetRequestHandler, specifically GroupBy in the 'transaction_number'. Instead of counting each transaction number using 'quantity_of_po', it now sums up all the quantities per group and represents it as 'item_count'. This is done to give a clearer overview of the total quantity per transaction, streamlining the process of managing assets.