Add AddingPrController route and modify AssetApprovalHandler function
This update introduces the `AddingPrController` route in the `api.php` file, allowing clients to create, retrieve, update, and delete Purchase Requests (PRs). We also made changes to the `transformIndexApproval` function in the `AssetApprovalHandler` trait. Now, this function takes an additional parameter `$transactionNumbers` and returns a quantity field calculated based on these transaction numbers. These modifications were necessary to compute and provide the total number of items in each transaction for the Asset Approval process.