Update Asset Request/Approval Processing and Add New Functionality
Updated instances of 'Declined' status to 'Returned' across various classes, for better terminology accuracy. Returned provides a clearer description of the asset request stage.

Implemented new moveData() function in AssetRequestController.php to take data from the RequestContainer and move it to the AssetRequest. This allows a more exible request flow and quicker form resubmissions. This required some edits for request handling traits and routes.

In CreateAssetApprovalRequest.php, added type declaration to improve code robustness and maintainability. Also, commented out (might be unnecessary) method in AssetRequest.php for future review.

With these changes, the user experience is expected to improve and the error proneness of code is reduced.