"Added new files for Request Container functionality
Implemented Request Container functionality that holds, validates, and manipulates requests' data before moving it to the Asset Request. This feature allows for things like validation, attachment storage, and more before the request becomes an actual Asset Request. The commit includes model, filters, migrations, and controller for the functionality and updated HTTP Requests accordingly. Also, a SubCapex filter was added for improved compatibility with the SubCapex functionality."