Refactoring AssetApprovalController: Removed TODO comment
The revisions made to AssetApprovalController.php involve the removal of an outdated TODO comment within the 'Return' case statement. The original comment implied that the term needed to be changed to 'Return', however, the term was already updated in a previous commit, making the comment redundant. This refactoring was made to reduce confusion and maintain the cleanliness of the codebase.