Update approval message and upgrade Laravel Excel version
In RequestContainerController, the approval message has been updated for better clarity. Now the message always indicates the approver as Approver 1 instead of dynamically assigning based on requester layer, to avoid confusion.

The Laravel Excel package in the composer.lock file has been upgraded from version 3.1.48 to 3.1.50. This change is to incorporate the benefits from the latest updates to the package.