Refactor request handling and validation rules
This commit aims to simplify the handling of input requests and validation rules in UpdateRequestContainerRequest, RequestContainerController, and CreateRequestContainerRequest. As a result, the application will handle requests more efficiently.

Changes are primarily made for the following reasons:
1. We no longer specify `'id'` after every field name in the request rules. This simplifies the code by removing redundant array notation. Now, the full field object is sent instead of only its 'id' property.
2. Commented out the unnecessary `department.company.company_id` rules, since 'company_id' is not directly under 'department_id' attribute in the request payload.

The changes will lead to a cleaner and better maintainable code because the validation rules now align with the request payload. Furthermore, the application performance will improve because of fewer nested array traversals.