"Improve asset request validation and asset approval handling
The cellphone number rule in the CreateAssetRequestRequest validation class was modified to be nullable without being numeric, providing greater flexibility for international formats. Function name in ApprovedRequestRepository.php was changed for better reflect its functionality. Improvements also made to asset resubmission logic in the ApprovedRequestRepository to restrict resubmissions to only returned requests, exemplifying defensive programming. New route added to 'show-by-id' on AssetRequestController. A transformation method for a single request item was added to AssetRequestHandler to get detailed data on an individual request. Lastly, adjustments in AssetApprovalController now clarify the usage of approval ids and remarks for approving or returning requests."