Fix remarks null value issue in ApprovedRequestRepository
The null coalescing operator was added to address potential null value issues in the 'remarks' field. Previously, this could cause errors when the status is 'Returned' and there are no remarks attached to the asset request. Now, the field is set to null in these scenarios to avoid any problems.