Add approver viewing feature
This commit modifies the AssetRequestController to allow approvers to view all asset requests tied to a specific transaction number, not just those they requested. Moreover, it refactors the way the "last layer" value is obtained in AssetRequestHandler, using the 'max' method instead of 'last' to get the highest layer number. This ensures a more accurate value if multiple layers have the same statuses. Finally, the ApprovedRequestRepository is explicitly defined in the AssetApprovalController; this enhancement enforces strong typing.