Refactor AssetRequestController and clean-up code
The AssetRequestController has been refactored to improve code quality and readability. Unnecessary blocks of code were removed as part of this process. This included some blocks of code that were commented out and not being used. The order of the code was also adjusted to enhance logical flow. Overall, the changes serve to make the code more manageable and easier to understand.